style: align to CR comments
